Modify GreatCircle hashCode calculation
This commit uses the same approach we've used in Vector2D and Vector3D
to calculate hash code value for the GreatCircle class instances.

Previously, GreatCircle.hashCode was using the Objects.hash(Object ...)
wrapper method. In its turn, it has called the Arrays.hashCode(Object[])
method.
Unfortunately, this does not prevent hash collisions calculated for the
GreatCircle instances. Not investigated the cause yet. As a guess, this
might be related to that the Precision.DoubleEquivalence's hashCode()
method is not overridden.