GEOMETRY-110: add ConvexHull3D using Quickhull algorithm
* GEOMETRY-110

Add ConvexHull3D class and tests.

* GEOMETRY-110

Return only an unmodifiable collection of facets.

* GEOMETRY-110

Test for a large number of points.

* Checkstyle Fix.

* GEOMETRY-110

Points are added to candidates again instead of a new point set.

* GEOMETRY-110

Only candidates points are distributed.

* GEOMETRY-110

Add Simplex type

* GEOMETRY-110

Create simplex while appending and cache it for later.

* GEOMETRY-110

Build simplex incrementially while appending points.

* GEOMETRY-110

Return facets as list and filter out duplicates in vertices.

* GEOMETRY-110

Returned collections have to be unmodifiable in public methods.

* GEOMETRY-110

Rename method and calculate maximum in a loop.

* GEOMETRY-110

Declare variable outside of loop.

* GEOMETRY-110

Rename methods.

* GEOMETRY-110

Replace bounds by array.

* GEOMETRY-110

Add check for appending a collection of points

* [GEOMETRY-110]

Delete hasOusidePoints and add exclusion for facets to spotbugs.

* [GEOMETRY-110]

Move isInside method into Facet class and reverse and rename it to isOutside

* [GEOMETRY-110]

Replace vertexToFacetMap with edgeMap

* [Geometry-110]

Change loop for creating edges.

* [GEOMETRY-110]

Facets are all located through edge operations. The edges are oriented in such a way, that neighbors share an edge in inverse orientation.

* [GEOMETRY-110]

Delete unnecessary debug code.

* [GEOMETRY-110]

Edges are oriented and lookup is done via an edgeMap. Each oriented edge is unique in association with a facet. Each facet stores the maximum offset and the associated point of outside points.

* [GEOMETRY-110]

- Rename Edge fields.
- Improve testing and assert all outputs of the algorithm.

* [GEOMETRY-110]

Check isDegenerate for all non-degenerate hulls.

* [GEOMETRY-110]

Even though the defined region is the same the number of facets can vary depending on the execution order of the algorithm.

---------

Co-authored-by: agoß <agoß@Goß>
Co-authored-by: Andreas Goss <agoss@itemis.com>